pyproject.toml
[project]
name = "book-recommendations"
version = "1.0.0"
description = "Book recommendation engine combining traditional ML with OpenAI"
authors = [
    { name = "Tadeas Fort", email = "business@tadeasfort.com" }
]
dependencies = [
    "pandas>=2.2.3",
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
    "numpy>=2.2.3",
    "scikit-learn>=1.6.1",
    "sqlmodel>=0.0.22",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "psycopg2-binary>=2.9.10",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "GPL-3.0" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.rye]
managed = true
dev-dependencies = [
]

[project.scripts]
book-recommendations = "book_recommendations.main:app"
alembic-init = "scripts.alembic-init:main"
alembic-migrate = "scripts.alembic-migrate:main"
alembic-push = "scripts.alembic-push:main"

[tool.rye.scripts]
init-db = { cmd = "./scripts/alembic-init.sh" }
migrate = { cmd = "./scripts/alembic-migrate.sh" }
push = { cmd = "./scripts/alembic-push.sh" }
